Completo el traslado de la accion
